Added cast to quiet compiler.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7916 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/std_pred_xsb_i.h
@@ -385,7 +385,7 @@ inline static xsbBool atom_to_list(CTXTdeclc int call_type)
 	  //	  printf("Allocated namebuf: %p, %d\n",atomnameaddr,atomnamelen);
 	}
 	//	if (CURRENT_CHARSET == UTF_8) 
-	  atomname = utf8_codepoint_to_str(c, atomname); /* nfz */
+	atomname = utf8_codepoint_to_str((int)c, atomname); /* nfz */
 	//	else *atomname++ = (char)c;  
 	term2 = get_list_tail(term2);
       } else {
